refactor: modularize AssetGrid and enhance styles for improved usability
- Extracted panels into separate components (AssetFoldersPanel and AssetFilesPanel) for better organization and maintainability.
- Introduced a custom hook (useClickOutsideDeselect) to streamline deselection logic when clicking outside selected assets.
- Updated AssetGrid styles to utilize a new assetGridStyles module, enhancing layout consistency and responsiveness.
- Improved README formatting for clarity and added installation notes for PyTorch with specific CUDA versions.
