+++ web/src/components/assets/panels/AssetFoldersPanel.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import FolderList from "../FolderList";
+
+const AssetFoldersPanel: React.FC = () => {
+  return (
+    <div
+      style={{
+        height: "100%",
+        overflowY: "auto",
+        overflowX: "hidden"
+      }}
+    >
+      <FolderList isHorizontal={false} />
+    </div>
+  );
+};
+
+export default AssetFoldersPanel;
